+++ magda-registry-api/src/test/scala/au/csiro/data61/magda/registry/WithDefaultPolicyRecordsServiceAuthSpec.scala
@@ -27,6 +27,7 @@ class WithDefaultPolicyRecordsServiceAuthSpec
        |authorization.skip = false
        |authorization.skipOpaQuery = false
        |opa.recordPolicyId = "default.policy"
+       |akka.loglevel = ERROR
     """.stripMargin
 
   describe("with default policy set") {
@@ -42,16 +43,23 @@ class WithDefaultPolicyRecordsServiceAuthSpec
           it(
             "allows access to an aspect-less record if default policy resolves to unconditionally allow access"
           ) { param =>
-            val recordId = "foo"
-            addRecord(param, Record(recordId, "foo", Map(), None))
-            expectOpaQueryForPolicy(
+            //   val recordId = "foo"
+            //   addRecord(param, Record(recordId, "foo", Map(), None))
+            //   expectOpaQueryForPolicy(
+            //     param,
+            //     "default.policy.read",
+            //     """{
+            //   "result": {
+            //       "queries": [[]]
+            //   }
+            // }"""
+            //   )
+
+            setupRecord(
               param,
+              None,
               "default.policy.read",
-              """{
-            "result": {
-                "queries": [[]]
-            }
-          }"""
+              policyResponseForUnconditionallyAllowed
             )
 
             Get(s"/v0/records/foo") ~> addTenantIdHeader(
@@ -68,12 +76,18 @@ class WithDefaultPolicyRecordsServiceAuthSpec
           it(
             "disallows access to an aspect-less record if default policy resolves to unconditionally disallow access"
           ) { param =>
-            val recordId = "foo"
+            //   val recordId = "foo"
 
-            addRecord(param, Record(recordId, "foo", Map(), None))
-            expectOpaQueryForPolicy(param, "default.policy.read", """{
-            "result": {}
-          }""")
+            //   addRecord(param, Record(recordId, "foo", Map(), None))
+            //   expectOpaQueryForPolicy(param, "default.policy.read", """{
+            //   "result": {}
+            // }""")
+            setupRecord(
+              param,
+              None,
+              "default.policy.read",
+              policyResponseForUnconditionallyDisallowed
+            )
 
             Get(s"/v0/records/foo") ~> addTenantIdHeader(
               TENANT_1
@@ -84,28 +98,20 @@ class WithDefaultPolicyRecordsServiceAuthSpec
 
           describe("based on the value in an aspect") {
             it(
-              "allows access to an record if policy resolves true"
+              "allows access to a record if policy resolves true"
             ) { param =>
               addExampleAspectDef(param)
-              val recordId = "foo"
-              addRecord(
-                param,
-                Record(
-                  recordId,
-                  "foo",
-                  Map(
-                    "stringExample" -> JsObject(
-                      "nested" -> JsObject("public" -> JsString("true"))
-                    )
-                  ),
-                  None
-                )
-              )
 
-              expectOpaQueryForPolicy(
+              setupRecord(
                 param,
+                None,
                 "default.policy.read",
-                policyResponseForStringExampleAspect
+                policyResponseForStringExampleAspect,
+                Map(
+                  "stringExample" -> JsObject(
+                    "nested" -> JsObject("public" -> JsString("true"))
+                  )
+                )
               )
 
               Get(s"/v0/records/foo") ~> addTenantIdHeader(
@@ -116,27 +122,19 @@ class WithDefaultPolicyRecordsServiceAuthSpec
             }
 
             it(
-              "denies access to an record if policy resolves false"
+              "denies access to a record if policy resolves false"
             ) { param =>
               addExampleAspectDef(param)
-              val recordId = "foo"
-              addRecord(
+              setupRecord(
                 param,
-                Record(
-                  recordId,
-                  "foo",
-                  Map(
-                    "stringExample" -> JsObject(
-                      "nested" -> JsObject("public" -> JsString("false"))
-                    )
-                  ),
-                  None
-                )
-              )
-              expectOpaQueryForPolicy(
-                param,
+                None,
                 "default.policy.read",
-                policyResponseForStringExampleAspect
+                policyResponseForStringExampleAspect,
+                Map(
+                  "stringExample" -> JsObject(
+                    "nested" -> JsObject("public" -> JsString("false"))
+                  )
+                )
               )
 
               Get(s"/v0/records/foo") ~> addTenantIdHeader(
@@ -147,21 +145,12 @@ class WithDefaultPolicyRecordsServiceAuthSpec
             }
 
             it(
-              "denies access to an record if required aspect is not present"
+              "denies access to a record if required aspect is not present"
             ) { param =>
               addExampleAspectDef(param)
-              val recordId = "foo"
-              addRecord(
+              setupRecord(
                 param,
-                Record(
-                  recordId,
-                  "foo",
-                  Map(),
-                  None
-                )
-              )
-              expectOpaQueryForPolicy(
-                param,
+                None,
                 "default.policy.read",
                 policyResponseForStringExampleAspect
               )
@@ -207,16 +196,11 @@ class WithDefaultPolicyRecordsServiceAuthSpec
           it(
             "allows access to an aspect-less record if default policy resolves to unconditionally allow access"
           ) { param =>
-            val recordId = "foo"
-            addRecord(param, Record(recordId, "foo", Map(), None))
-            expectOpaQueryForPolicy(
+            setupRecord(
               param,
+              None,
               "default.policy.read",
-              """{
-              "result": {
-                  "queries": [[]]
-              }
-            }"""
+              policyResponseForUnconditionallyAllowed
             )
 
             Get(s"/v0/records/summary/foo") ~> addTenantIdHeader(
@@ -232,12 +216,12 @@ class WithDefaultPolicyRecordsServiceAuthSpec
           it(
             "disallows access to an aspect-less record if default policy resolves to unconditionally disallow access"
           ) { param =>
-            val recordId = "foo"
-
-            addRecord(param, Record(recordId, "foo", Map(), None))
-            expectOpaQueryForPolicy(param, "default.policy.read", """{
-              "result": {}
-            }""")
+            setupRecord(
+              param,
+              None,
+              "default.policy.read",
+              policyResponseForUnconditionallyDisallowed
+            )
 
             Get(s"/v0/records/summary/foo") ~> addTenantIdHeader(
               TENANT_1
@@ -248,28 +232,20 @@ class WithDefaultPolicyRecordsServiceAuthSpec
 
           describe("based on the value in an aspect") {
             it(
-              "allows access to an record if policy resolves true"
+              "allows access to a record if policy resolves true"
             ) { param =>
               addExampleAspectDef(param)
-              val recordId = "foo"
-              addRecord(
-                param,
-                Record(
-                  recordId,
-                  "foo",
-                  Map(
-                    "stringExample" -> JsObject(
-                      "nested" -> JsObject("public" -> JsString("true"))
-                    )
-                  ),
-                  None
-                )
-              )
 
-              expectOpaQueryForPolicy(
+              setupRecord(
                 param,
+                None,
                 "default.policy.read",
-                policyResponseForStringExampleAspect
+                policyResponseForStringExampleAspect,
+                Map(
+                  "stringExample" -> JsObject(
+                    "nested" -> JsObject("public" -> JsString("true"))
+                  )
+                )
               )
 
               Get(s"/v0/records/summary/foo") ~> addTenantIdHeader(
@@ -282,27 +258,19 @@ class WithDefaultPolicyRecordsServiceAuthSpec
             }
 
             it(
-              "denies access to an record if policy resolves false"
+              "denies access to a record if policy resolves false"
             ) { param =>
               addExampleAspectDef(param)
-              val recordId = "foo"
-              addRecord(
+              setupRecord(
                 param,
-                Record(
-                  recordId,
-                  "foo",
-                  Map(
-                    "stringExample" -> JsObject(
-                      "nested" -> JsObject("public" -> JsString("false"))
-                    )
-                  ),
-                  None
-                )
-              )
-              expectOpaQueryForPolicy(
-                param,
+                None,
                 "default.policy.read",
-                policyResponseForStringExampleAspect
+                policyResponseForStringExampleAspect,
+                Map(
+                  "stringExample" -> JsObject(
+                    "nested" -> JsObject("public" -> JsString("false"))
+                  )
+                )
               )
 
               Get(s"/v0/records/summary/foo") ~> addTenantIdHeader(
@@ -313,26 +281,262 @@ class WithDefaultPolicyRecordsServiceAuthSpec
             }
 
             it(
-              "denies access to an record if required aspect is not present"
+              "denies access to a record if required aspect is not present"
             ) { param =>
               addExampleAspectDef(param)
-              val recordId = "foo"
-              addRecord(
+              setupRecord(
                 param,
-                Record(
-                  recordId,
-                  "foo",
-                  Map(),
-                  None
-                )
-              )
-              expectOpaQueryForPolicy(
-                param,
+                None,
                 "default.policy.read",
                 policyResponseForStringExampleAspect
               )
 
-              Get(s"/v0/records/foo") ~> addTenantIdHeader(
+              Get(s"/v0/records/summary/foo") ~> addTenantIdHeader(
+                TENANT_1
+              ) ~> param.api(Full).routes ~> check {
+                status shouldEqual StatusCodes.NotFound
+              }
+            }
+          }
+        }
+      }
+
+      describe("for record history") {
+        describe("with a policy set on the record") {
+          commonRecordHistoryTests(Some("default.policy"), true)
+        }
+
+        describe("with no policy set on the record") {
+          commonRecordHistoryTests(Some("default.policy"), false)
+
+          it(
+            "shows events from an aspect-less record if default policy resolves to unconditionally allow access"
+          ) { param =>
+            setupRecord(
+              param,
+              None,
+              "default.policy.read",
+              policyResponseForUnconditionallyAllowed
+            )
+
+            Get(s"/v0/records/foo/history") ~> addTenantIdHeader(
+              TENANT_1
+            ) ~> param.api(Full).routes ~> check {
+              status shouldEqual StatusCodes.OK
+              val resRecord = responseAs[EventsPage]
+
+              resRecord.events.head.data
+                .fields("recordId") shouldEqual JsString("foo")
+            }
+          }
+
+          it(
+            "doesn't show events from an aspect-less record if default policy resolves to unconditionally disallow access"
+          ) { param =>
+            setupRecord(
+              param,
+              None,
+              "default.policy.read",
+              policyResponseForUnconditionallyDisallowed
+            )
+
+            Get(s"/v0/records/foo/history") ~> addTenantIdHeader(
+              TENANT_1
+            ) ~> param.api(Full).routes ~> check {
+              status shouldEqual StatusCodes.OK
+              val resRecord = responseAs[EventsPage]
+              resRecord.events.length shouldBe 0
+            }
+          }
+
+          describe("based on the value in an aspect") {
+            it(
+              "shows events from a record if policy resolves true"
+            ) { param =>
+              addExampleAspectDef(param)
+
+              setupRecord(
+                param,
+                None,
+                "default.policy.read",
+                policyResponseForStringExampleAspect,
+                Map(
+                  "stringExample" -> JsObject(
+                    "nested" -> JsObject("public" -> JsString("true"))
+                  )
+                )
+              )
+
+              Get(s"/v0/records/foo/history") ~> addTenantIdHeader(
+                TENANT_1
+              ) ~> param.api(Full).routes ~> check {
+                status shouldEqual StatusCodes.OK
+                val resRecord = responseAs[EventsPage]
+                resRecord.events.head.data
+                  .fields("recordId") shouldEqual JsString("foo")
+              }
+            }
+
+            it(
+              "shows no events from a record if policy resolves false"
+            ) { param =>
+              addExampleAspectDef(param)
+              setupRecord(
+                param,
+                None,
+                "default.policy.read",
+                policyResponseForStringExampleAspect,
+                Map(
+                  "stringExample" -> JsObject(
+                    "nested" -> JsObject("public" -> JsString("false"))
+                  )
+                )
+              )
+
+              Get(s"/v0/records/foo/history") ~> addTenantIdHeader(
+                TENANT_1
+              ) ~> param.api(Full).routes ~> check {
+                status shouldEqual StatusCodes.OK
+                val resRecord = responseAs[EventsPage]
+                resRecord.events.length shouldBe 0
+              }
+            }
+
+            it(
+              "shows no events from a record if required aspect is not present"
+            ) { param =>
+              addExampleAspectDef(param)
+              setupRecord(
+                param,
+                None,
+                "default.policy.read",
+                policyResponseForStringExampleAspect
+              )
+
+              Get(s"/v0/records/foo/history") ~> addTenantIdHeader(
+                TENANT_1
+              ) ~> param.api(Full).routes ~> check {
+                status shouldEqual StatusCodes.OK
+                val resRecord = responseAs[EventsPage]
+                resRecord.events.length shouldBe 0
+              }
+            }
+          }
+        }
+      }
+
+      describe("for record version (time-travel)") {
+        describe("with a policy set on the record") {
+          commonVersionTests(Some("default.policy"), true)
+        }
+
+        describe("with no policy set on the record") {
+          commonVersionTests(Some("default.policy"), false)
+
+          it(
+            "shows an aspect-less record if default policy resolves to unconditionally allow access"
+          ) { param =>
+            setupRecord(
+              param,
+              None,
+              "default.policy.read",
+              policyResponseForUnconditionallyAllowed
+            )
+
+            Get(s"/v0/records/foo/history/${Integer.MAX_VALUE}") ~> addTenantIdHeader(
+              TENANT_1
+            ) ~> param.api(Full).routes ~> check {
+              status shouldEqual StatusCodes.OK
+              val resRecord = responseAs[Record]
+
+              resRecord.id shouldEqual "foo"
+            }
+          }
+
+          it(
+            "returns 404 for a record if default policy resolves to unconditionally disallow access"
+          ) { param =>
+            setupRecord(
+              param,
+              None,
+              "default.policy.read",
+              policyResponseForUnconditionallyDisallowed
+            )
+
+            Get(s"/v0/records/foo/history/${Integer.MAX_VALUE}") ~> addTenantIdHeader(
+              TENANT_1
+            ) ~> param.api(Full).routes ~> check {
+              status shouldEqual StatusCodes.NotFound
+            }
+          }
+
+          describe("based on the value in an aspect") {
+            it(
+              "shows a record if default policy resolves true"
+            ) { param =>
+              addExampleAspectDef(param)
+
+              setupRecord(
+                param,
+                None,
+                "default.policy.read",
+                policyResponseForStringExampleAspect,
+                Map(
+                  "stringExample" -> JsObject(
+                    "nested" -> JsObject("public" -> JsString("true"))
+                  )
+                )
+              )
+
+              Get(s"/v0/records/foo/history/${Integer.MAX_VALUE}") ~> addTenantIdHeader(
+                TENANT_1
+              ) ~> param.api(Full).routes ~> check {
+                status shouldEqual StatusCodes.OK
+                val resRecord = responseAs[Record]
+                resRecord.id shouldEqual "foo"
+                resRecord.aspects shouldEqual Map(
+                  "stringExample" -> JsObject(
+                    "nested" -> JsObject("public" -> JsString("true"))
+                  )
+                )
+              }
+            }
+
+            it(
+              "returns 404 for a record if default policy resolves false"
+            ) { param =>
+              addExampleAspectDef(param)
+              setupRecord(
+                param,
+                None,
+                "default.policy.read",
+                policyResponseForStringExampleAspect,
+                Map(
+                  "stringExample" -> JsObject(
+                    "nested" -> JsObject("public" -> JsString("false"))
+                  )
+                )
+              )
+
+              Get(s"/v0/records/foo/history/${Integer.MAX_VALUE}") ~> addTenantIdHeader(
+                TENANT_1
+              ) ~> param.api(Full).routes ~> check {
+                status shouldEqual StatusCodes.NotFound
+              }
+            }
+
+            it(
+              "returns 404 for a record if required aspect for the default policy is not present"
+            ) { param =>
+              addExampleAspectDef(param)
+              setupRecord(
+                param,
+                None,
+                "default.policy.read",
+                policyResponseForStringExampleAspect
+              )
+
+              Get(s"/v0/records/foo/history/${Integer.MAX_VALUE}") ~> addTenantIdHeader(
                 TENANT_1
               ) ~> param.api(Full).routes ~> check {
                 status shouldEqual StatusCodes.NotFound