+++ magda-web-server/src/index.ts
@@ -110,6 +110,11 @@ const argv = yargs
             "The base URL of the correspondence api.  If not specified, the URL is built from the apiBaseUrl.",
         type: "string"
     })
+    .option("openfaasBaseUrl", {
+        describe:
+            "The base URL of the openfaas gateway.  If not specified, the URL is built from the apiBaseUrl.",
+        type: "string"
+    })
     .option("fallbackUrl", {
         describe:
             "An older system to fall back to - this url will be shown in a banner that says 'you can still go back to old site'.",
@@ -273,6 +278,13 @@ const webServerConfig = {
             new URI(apiBaseUrl)
                 .segment("v0")
                 .segment("correspondence")
+                .toString()
+    ),
+    openfaasBaseUrl: addTrailingSlash(
+        argv.openfaasBaseUrl ||
+            new URI(apiBaseUrl)
+                .segment("v0")
+                .segment("openfaas")
                 .toString()
     ),
     fallbackUrl: argv.fallbackUrl,