+++ CHANGES.md
@@ -16,6 +16,7 @@ General:
 -   Move [magda-preview-map](https://github.com/magda-io/magda-preview-map) out of the core repo
 -   Upgraded everything cert-manager related to work with v0.13
 -   Update build & run document to provide more information regarding running local build connector & minion docker images
+-   New: Extract metadata from data portal URLs
 
 UI:
 
@@ -50,6 +51,7 @@ Storage:
 -   Support multipart upload
 -   Fixed: minio chart will not be deployed if storage-api is not turned on
 -   Make MinIO a dependency of storage's helm chart
+-   Make minio, by default, deploy with `Recreate` strategy
 
 Gateway:
 
@@ -65,6 +67,8 @@ Authorization:
 -   Added authorization inside links with dereferencing on or off, for the `/records` endpoint
 -   Added per-record authorization around the `/records/summary/<recordid>` endpoint matching the `/records/<id>` one.
 -   Added per-record authorization around the `/records/summary` endpoint matching the `/records` one.
+-   Added per-record authorization around the `/records/<recordid>/history` endpoint
+-   Added per-record authorization around the `/records/<recordid>/history/<eventId>` endpoint
 
 Others: